Split GroupTest into multiple files
Part of: #34494
